Make `node-gyp` a development dependency
### Why?

According to the Node
[documentation](https://nodejs.org/docs/latest-v14.x/api/addons.html):

> A version of the `node-gyp` utility is bundled and distributed with
> Node.js as part of npm.

If this is the case, why are we forcing library users to download a
separate version?

If we make `node-gyp` a development dependency then we radically reduce
the number of transitive dependencies that library users must use.
Here's the new situation:

```
success Saved 4 new dependencies.
info Direct dependencies
└─ spookyhash@0.2.0
info All dependencies
├─ bindings@1.5.0
├─ file-uri-to-path@1.0.0
├─ node-addon-api@3.1.0
└─ spookyhash@0.2.0
```

We used to save **95** new dependencies!

### What?

* move `node-gyp` from `dependencies` to `devDependencies`