Switched tests to prefer `Buffer` for seed values
### Why?

To fix #2 we'll drop `BigInt` support when building the library on early
versions of Node.js. We'd still like to run tests on those versions, so
will need to make the test cases conditional on the version. To make
this easier we should make most of the tests use `Buffer` and group the
`BigInt` tests together in each test file.

### What?

* change tests to prefer `Buffer` to `BigInt`
* add separate sets of seed tests for each type